requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-sso==0.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-sso==0.17.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
coverage==7.6.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-sso==0.17.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
africastalking==1.2.9
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coverage==7.6.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-sso==0.17.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
schema==0.7.7
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
